% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-checks.R
\name{check_session}
\alias{check_session}
\title{Check SAS session}
\usage{
check_session(call = rlang::caller_env())
}
\description{
Checks if a SAS session currently exists. If the SAS session has terminated
during the session, \code{check_session()} will not detect it.
}
\details{
Use \code{execute_if_connection_active()} for any function that relies on a SAS
connection to catch inactive sessions.
}
\keyword{internal}
