% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{chk_connection}
\alias{chk_connection}
\title{Check SAS connection}
\usage{
chk_connection()
}
\description{
Checks if a SAS session currently exists. If the SAS session has terminated
during the session, \code{chk_connection()} will not detect it.
}
\details{
Use \code{execute_if_connection_active()} for any function that relies on a SAS
connection to catch inactive sessions.
}
\keyword{internal}
