% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-checks.R
\name{chk_session}
\alias{chk_session}
\title{Check SAS session}
\usage{
chk_session()
}
\description{
Checks if a SAS session currently exists. If the SAS session has terminated
during the session, \code{chk_session()} will not detect it.
}
\details{
Use \code{execute_if_connection_active()} for any function that relies on a SAS
connection to catch inactive sessions.
}
\keyword{internal}
