% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-checks.R
\name{execute_if_connection_active}
\alias{execute_if_connection_active}
\title{Execute SASPy function if session is active}
\usage{
execute_if_connection_active(code)
}
\description{
Executes the code passed to it if the connection is active and provides a
more helpful error message if no connection is active.
}
\details{
The SAS connection is asynchronous so it can become inactive without the user
knowing. When the connection is inactive and an action is preformed it will
check the connection and raise an error.
}
\keyword{internal}
