% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{r_to_sas}
\alias{r_to_sas}
\title{Convert R table to SAS}
\usage{
r_to_sas(x, table_name, libref = "WORK")
}
\value{
No return value.
}
\description{
Converts R table into a table in the current SAS session. Translation errors
may occur.
}
