% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sas_connect}
\alias{sas_connect}
\title{Establish SAS session}
\usage{
sas_connect(config)
}
\arguments{
\item{config}{Configuration to use from the SAS_config_names list in
\code{sascfg_personal.py}.}
}
\value{
No return value.
}
\description{
Starts a SAS session. This is required before doing anything!
}
