% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sas_download}
\alias{sas_download}
\title{Download a file from SAS}
\usage{
sas_download(sas_path, local_path)
}
\arguments{
\item{sas_path}{Path of file on remote SAS server to be download}

\item{local_path}{Path to upload SAS file to on local machine.}
}
\value{
No return value.
}
\description{
Downloads a file to the remote SAS server.
}
