% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-copy.R
\name{sas_file_copy}
\alias{sas_file_copy}
\title{Copy a file on SAS}
\usage{
sas_file_copy(from_path, to_path)
}
\arguments{
\item{from_path}{string; Path of file on remote SAS server to be copied.}

\item{to_path}{string; Path of file on remote SAS server to copy to.}
}
\value{
\code{logical}; value indicating if the operation succeeded.
}
\description{
Copies a file on the remote SAS server. Is analogous to
\code{file.copy()}, but for the remote SAS server.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# connect to SAS
sas_connect()

# create a file and upload it to SAS
tempfile_path <- tempfile(fileext = ".sas")
tempfile_basename <- basename(tempfile_path)
tempfile_basename_copy <- sub("\\\\.sas$", "_copy.sas", tempfile_basename)
cat("PROC MEANS DATA = sashelp.cars;RUN;", file = tempfile_path)
sas_file_upload(local_path = tempfile_path, sas_path = paste0("~/", tempfile_basename))

# copy file on SAS
sas_file_copy(paste0("~/", tempfile_basename), paste0("~/", tempfile_basename_copy))
\dontshow{\}) # examplesIf}
}
\seealso{
Other file management functions: 
\code{\link{sas_file_download}()},
\code{\link{sas_file_exists}()},
\code{\link{sas_file_remove}()},
\code{\link{sas_file_upload}()},
\code{\link{sas_list}()}
}
\concept{file management functions}
