% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-download.R
\name{sas_file_download}
\alias{sas_file_download}
\title{Download a file from SAS}
\usage{
sas_file_download(sas_path, local_path)
}
\arguments{
\item{sas_path}{string; Path of file on remote SAS server to be download}

\item{local_path}{string; Path to upload SAS file to on local machine.}
}
\value{
\code{logical}; value indicating if the operation succeeded.
}
\description{
Downloads a file to the remote SAS server.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# connect to SAS
sas_connect()

# create a file and upload it to SAS
tempfile_path <- tempfile(fileext = ".sas")
tempfile_basename <- basename(tempfile_path)
tempfile_path_copy <- sub("\\\\.sas$", "_copy.sas", tempfile_path)
cat("PROC MEANS DATA = sashelp.cars;RUN;", file = tempfile_path)
sas_file_upload(local_path = tempfile_path, sas_path = paste0("~/", tempfile_basename))

# download file from SAS
sas_file_download(paste0("~/", tempfile_basename), tempfile_path_copy)
\dontshow{\}) # examplesIf}
}
\seealso{
Other file management functions: 
\code{\link{sas_file_copy}()},
\code{\link{sas_file_exists}()},
\code{\link{sas_file_remove}()},
\code{\link{sas_file_upload}()},
\code{\link{sas_list}()}
}
\concept{file management functions}
