% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sas_get_session}
\alias{sas_get_session}
\title{Get current SAS session}
\usage{
sas_get_session()
}
\value{
Current SAS session.
}
\description{
Gets current SAS session so that you can use functions not yet implemented.
Can also be useful for testing or using Python.
}
\examples{
\dontrun{
sas_connect()
sas_get_session()
}
}
