% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sas_list}
\alias{sas_list}
\title{List contents of a SAS directory}
\usage{
sas_list(path)
}
\arguments{
\item{path}{Path of directory on remote SAS server to list the contents of.}
}
\value{
No return value.
}
\description{
Lists the files or directories of a directory within the remote SAS server.
}
\examples{
\dontrun{
sas_connect()
sas_list(".")
}
}
