% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{sas_remove}
\alias{sas_remove}
\title{Delete a file or directory from SAS}
\usage{
sas_remove(path)
}
\arguments{
\item{path}{Path of file on remote SAS server to be deleted.}
}
\value{
No return value.
}
\description{
Deletes a file or directory from the remote SAS server.
}
