% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-addin.R
\name{sas_run_selected}
\alias{sas_run_selected}
\title{Execute selected SAS code}
\usage{
sas_run_selected()
}
\value{
No return value.
}
\description{
Execute selected SAS code in current session and render html output.

You can add the following to your Positron keybindings.json and run SAS
code by selecting it and pressing ctrl+shift+enter.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\{
    "key": "ctrl+shift+enter",
    "command": "workbench.action.executeCode.console",
    "when": "editorTextFocus",
    "args": \{
        "langId": "r",
        "code": "sasquatch::sas_run_selected()",
        "focus": true
    \}
\}
}\if{html}{\out{</div>}}
}
