% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{sas_to_r}
\alias{sas_to_r}
\title{Convert SAS table to R}
\usage{
sas_to_r(table_name, libref = "WORK")
}
\arguments{
\item{table_name}{Name of table in SAS.}

\item{libref}{Name of libref SAS table is stored within.}
}
\value{
R \code{data.frame}.
}
\description{
Converts table from current SAS session into a R \code{data.frame}. Translation
errors may occur.
}
