% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{sas_upload}
\alias{sas_upload}
\title{Upload a file to SAS}
\usage{
sas_upload(local_path, sas_path)
}
\arguments{
\item{local_path}{Path of file on local machine to be uploaded.}

\item{sas_path}{Path to upload local file to on the remote SAS server.}
}
\value{
No return value.
}
\description{
Uploads a file to the remote SAS server.
}
