% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto.R
\name{sas_use_quarto}
\alias{sas_use_quarto}
\title{Add SAS quarto template to specified directory}
\usage{
sas_use_quarto(path)
}
\arguments{
\item{path}{(Default: Current directory) Path to copy quarto template and
syntax highlighting XML into.}
}
\value{
No return value.
}
\description{
Adds SAS quarto template and syntax highlighting XML to a specified
directory.
}
