% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasquatch-package.R
\docType{package}
\name{sasquatch-package}
\alias{sasquatch}
\alias{sasquatch-package}
\title{sasquatch: Use 'SAS', R, and 'quarto' Together}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use R and 'SAS' within reproducible multilingual 'quarto' documents. Run 'SAS' code blocks interactively, send data back and forth between 'SAS' and R, and render 'SAS' output within quarto documents. 'SAS' connections are established through a combination of 'SASPy' and 'reticulate'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ryanzomorrodi.github.io/sasquatch/}
  \item \url{https://github.com/ryanzomorrodi/sasquatch}
  \item Report bugs at \url{https://github.com/ryanzomorrodi/sasquatch/issues}
}

}
\author{
\strong{Maintainer}: Ryan Zomorrodi \email{rzomor2@uic.edu} (\href{https://orcid.org/0009-0003-6417-5985}{ORCID}) [copyright holder]

}
\keyword{internal}
