% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto.R
\name{use_sasr_quarto}
\alias{use_sasr_quarto}
\title{Add sasr quarto template to current directory}
\usage{
use_sasr_quarto()
}
\value{
No return value.
}
\description{
Adds sasr quarto template and syntax highlighting XML to current directory.
}
