% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{r_to_sas}
\alias{r_to_sas}
\title{Convert R table to SAS}
\usage{
r_to_sas(x, table_name, libref = "WORK")
}
\arguments{
\item{x}{data.frame, tibble, data.table; R table.}

\item{table_name}{string; Name of table to be created in SAS.}

\item{libref}{string; Name of libref to store SAS table within.}
}
\value{
No return value.
}
\description{
Converts R table into a table in the current SAS session. R tables must only
have logical, integer, double, factor, character, POSIXct, or Date class
columns.
}
\details{
SAS only has two data types (numeric and character). Data types are converted
as follows:
\itemize{
\item logical -> numeric
\item integer -> numeric
\item double -> numeric
\item factor -> character
\item character -> character
\item POSIXct -> numeric (datetime)
\item Date -> numeric (date)
}
}
\examples{
\dontrun{
sas_connect()
r_to_sas(mtcars, "mtcars")
}
}
