% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sas_connect}
\alias{sas_connect}
\title{Establish SAS session}
\usage{
sas_connect(cfgname)
}
\arguments{
\item{cfgname}{string; Name of configuration to use from the SAS_config_names
list in \code{sascfg_personal.py}.}
}
\value{
No return value.
}
\description{
Starts a SAS session. This is required before doing anything!
}
\details{
All configurations are specified within the \code{sascfg_personal.py} file inside
the \code{SASPy} package. For more information about \code{SASPy} configuration, check
out the \href{https://sassoftware.github.io/saspy/configuration.html}{configuration documentation}
or \code{vignette("setting_up")}.
}
\examples{
\dontrun{
sas_connect(cfgname = "oda")
}
}
