% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{sas_disconnect}
\alias{sas_disconnect}
\title{Disconnect SAS session}
\usage{
sas_disconnect()
}
\value{
No return value.
}
\description{
Disconnects from the current SAS session.
}
\examples{
\dontrun{
sas_disconnect()
}
}
\seealso{
Other session management functions: 
\code{\link{sas_connect}()},
\code{\link{sas_get_session}()}
}
\concept{session management functions}
