% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{sas_file_download}
\alias{sas_file_download}
\title{Download a file from SAS}
\usage{
sas_file_download(sas_path, local_path)
}
\arguments{
\item{sas_path}{Path of file on remote SAS server to be download}

\item{local_path}{Path to upload SAS file to on local machine.}
}
\value{
Logical value indicating if the operation succeeded.
}
\description{
Downloads a file to the remote SAS server.
}
\examples{
\dontrun{
# connect to SAS
sas_connect()

# create a file and upload it to SAS
cat("PROC MEANS DATA = sashelp.cars;RUN;", file = "script.sas")
sas_file_upload(local_path = "script.sas", sas_path = "~/script.sas")

# download file from SAS
sas_file_download(sas_path = "~/script.sas", local_path = "script.sas")
}
}
