% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-remove.R
\name{sas_file_remove}
\alias{sas_file_remove}
\title{Delete a file or directory from SAS}
\usage{
sas_file_remove(path)
}
\arguments{
\item{path}{string; Path of file on remote SAS server to be deleted.}
}
\value{
\code{logical}; value indicating if the operation succeeded.
}
\description{
Deletes a file or directory from the remote SAS server. Is analogous to
\code{file.remove()}, but for the remote SAS server.
}
\examples{
\dontrun{
# connect to SAS
sas_connect()

# create a file and upload it to SAS
cat("PROC MEANS DATA = sashelp.cars;RUN;", file = "script.sas")
sas_file_upload(local_path = "script.sas", sas_path = "~/script.sas")

# remove file from SAS
sas_file_remove(sas_path = "~/script.sas")
}
}
