% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-upload.R
\name{sas_file_upload}
\alias{sas_file_upload}
\title{Upload a file to SAS}
\usage{
sas_file_upload(local_path, sas_path)
}
\arguments{
\item{local_path}{Path of file on local machine to be uploaded.}

\item{sas_path}{Path to upload local file to on the remote SAS server.}
}
\value{
Logical value indicating if the operation succeeded.
}
\description{
Uploads a file to the remote SAS server.
}
\examples{
\dontrun{
# connect to SAS
sas_connect()

# create a file to upload
cat("PROC MEANS DATA = sashelp.cars;RUN;", file = "script.sas")

# upload file
sas_file_upload(local_path = "script.sas", sas_path = "~/script.sas")
}
}
