% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_r.R
\name{sas_from_r}
\alias{sas_from_r}
\title{Convert R table to SAS}
\usage{
sas_from_r(x, table_name, libref = "WORK")
}
\arguments{
\item{x}{\code{data.frame}; R table.}

\item{table_name}{string; Name of table to be created in SAS.}

\item{libref}{string; Name of libref to store SAS table within.}
}
\value{
\code{data.frame}; \code{x}.
}
\description{
Converts R table into a table in the current SAS session. R tables must only
have logical, integer, double, factor, character, POSIXct, or Date class
columns.
}
\details{
SAS only has two data types (numeric and character). Data types are converted
as follows:
\itemize{
\item logical -> numeric
\item integer -> numeric
\item double -> numeric
\item factor -> character
\item character -> character
\item POSIXct -> numeric (datetime)
\item Date -> numeric (date)
}
}
\examples{
\dontrun{
sas_connect()
sas_from_r(mtcars, "mtcars")
}
}
\seealso{
\code{\link[=sas_to_r]{sas_to_r()}}
}
