% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{sas_list}
\alias{sas_list}
\title{List contents of a SAS directory}
\usage{
sas_list(path)
}
\arguments{
\item{path}{string; Path of directory on remote SAS server to list the contents of.}
}
\value{
\code{character} vector; File or directory names.
}
\description{
Lists the files or directories of a directory within the remote SAS server.
}
\examples{
\dontrun{
sas_connect()
sas_list(".")
}
}
\seealso{
Other file management functions: 
\code{\link{sas_file_copy}()},
\code{\link{sas_file_download}()},
\code{\link{sas_file_exists}()},
\code{\link{sas_file_remove}()},
\code{\link{sas_file_upload}()}
}
\concept{file management functions}
