% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sas_remove}
\alias{sas_remove}
\title{Delete a file or directory from SAS}
\usage{
sas_remove(path)
}
\arguments{
\item{path}{Path of file on remote SAS server to be deleted.}
}
\value{
No return value.
}
\description{
Deletes a file or directory from the remote SAS server.
}
\examples{
\dontrun{
sas_connect()

# upload document to SAS
quarto_file <- system.file("sasquatch.qmd", package = "sasquatch")
sas_upload(quarto_file, "sasquatch.qmd")

# remove document from SAS
sas_remove("sasquatch.qmd")
}
}
