% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{sas_run_file}
\alias{sas_run_file}
\title{Execute SAS file}
\usage{
sas_run_file(input_path, output_path, overwrite = FALSE)
}
\arguments{
\item{input_path}{Path of SAS file to run.}

\item{output_path}{Optional path to save html output to (log file will be
named the same).}

\item{overwrite}{Can output overwrite prior output?}
}
\value{
No return value.
}
\description{
Execute a SAS file and render html output or save output as html and log.
}
