% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-selected.R
\name{sas_run_selected}
\alias{sas_run_selected}
\title{Execute selected SAS code}
\usage{
sas_run_selected()
}
\value{
No return value.
}
\description{
Execute selected SAS code in current session and render html output as SAS
widget.
}
\examples{
\dontrun{
sas_connect()

# highlight something in the active editor of RStudio or Positron

sas_run_selected()
}
}
