% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-selected.R
\name{sas_run_selected}
\alias{sas_run_selected}
\title{Execute selected SAS code}
\usage{
sas_run_selected()
}
\value{
\code{htmlwidget}; HTML5 output.
}
\description{
Execute selected SAS code in current session and render html output as SAS
widget. See \code{vignette("sasquatch")} for more information on how to utilize
the addin within RStudio or Positron.
}
\examples{
\dontrun{
sas_connect()

# highlight something in the active editor of RStudio or Positron

sas_run_selected()
}
}
\seealso{
Other code execution functions: 
\code{\link{sas_run_file}()},
\code{\link{sas_run_string}()}
}
\concept{code execution functions}
