% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-string.R
\name{sas_run_string}
\alias{sas_run_string}
\title{Execute SAS code}
\usage{
sas_run_string(input)
}
\arguments{
\item{input}{String of SAS code input to run.}
}
\value{
No return value.
}
\description{
Execute SAS code in current session and render html output.
}
\examples{
\dontrun{
sas_connect()
sas_run_string("PROC MEANS DATA = sashelp.cars;\n RUN;")
}
}
