% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-string.R
\name{sas_run_string}
\alias{sas_run_string}
\title{Execute SAS code string}
\usage{
sas_run_string(input)
}
\arguments{
\item{input}{string; SAS code to run.}
}
\value{
\code{htmlwidget}; HTML5 output.
}
\description{
Execute SAS code in current session and render html output.
}
\examples{
\dontrun{
sas_connect()
sas_run_string("PROC MEANS DATA = sashelp.cars;RUN;")
}
}
\seealso{
Other code execution functions: 
\code{\link{sas_run_file}()},
\code{\link{sas_run_selected}()}
}
\concept{code execution functions}
