% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto.R
\name{sas_use_quarto}
\alias{sas_use_quarto}
\title{Add sasquatch quarto template to current directory}
\usage{
sas_use_quarto()
}
\value{
No return value.
}
\description{
Adds sasquatch quarto template and syntax highlighting XML to current directory.
}
